+++ extensions/terminal-suggest/src/completions/code-tunnel.ts
@@ -2,7 +2,12 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import code, { buildSubcommandWithCommonOptions, commonCLIOptions, commonOptions, extensionManagementOptions, troubleshootingOptions } from './code';
+import code, { commonCLIOptions, commonOptions, extensionManagementOptions, troubleshootingOptions } from './code';
+
+export const buildSubcommandWithCommonOptions = (command: any) => ({
+	...command,
+	options: [...(command.options || []), ...commonCLIOptions]
+});
 
 const commonAuthOptions: Fig.Option[] = [
 	{
@@ -38,7 +43,7 @@ const commonAuthOptions: Fig.Option[] = [
 	}
 ];
 
-export const tunnelSubcommands: Fig.Subcommand[] = [
+const tunnelSubcommands: Fig.Subcommand[] = [
 	{
 		name: 'prune',
 		description: 'Delete all servers which are currently not running',
@@ -254,6 +259,189 @@ const codeTunnelCompletionSpec: Fig.Spec = {
 		},
 	]
 };
+export const codeTunnelSubcommands = [{
+	name: 'serve-web',
+	description: 'Runs a local web version of Code - OSS',
+	options: [
+		{
+			name: '--host',
+			description: 'Host to listen on, defaults to \'localhost\'',
+			isRepeatable: true,
+			args: {
+				name: 'host',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--socket-path',
+			isRepeatable: true,
+			args: {
+				name: 'socket_path',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--port',
+			description: 'Port to listen on. If 0 is passed a random free port is picked',
+			isRepeatable: true,
+			args: {
+				name: 'port',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--connection-token',
+			description: 'A secret that must be included with all requests',
+			isRepeatable: true,
+			args: {
+				name: 'connection_token',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--connection-token-file',
+			description: 'A file containing a secret that must be included with all requests',
+			isRepeatable: true,
+			args: {
+				name: 'connection_token_file',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--server-base-path',
+			description: 'Specifies the path under which the web UI and the code server is provided',
+			isRepeatable: true,
+			args: {
+				name: 'server_base_path',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--server-data-dir',
+			description: 'Specifies the directory that server data is kept in',
+			isRepeatable: true,
+			args: {
+				name: 'server_data_dir',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--without-connection-token',
+			description: 'Run without a connection token. Only use this if the connection is secured by other means',
+		},
+		{
+			name: '--accept-server-license-terms',
+			description: 'If set, the user accepts the server license terms and the server will be started without a user prompt',
+		},
+		...commonCLIOptions,
+	]
+},
+{
+	name: 'tunnel',
+	description: 'Create a tunnel that\'s accessible on vscode.dev from anywhere.Run`code tunnel --help` for more usage info',
+	subcommands: tunnelSubcommands,
+	options: [
+		{
+			name: '--install-extension',
+			description: 'Requests that extensions be preloaded and installed on connecting servers',
+			isRepeatable: true,
+			args: {
+				name: 'install_extension',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--server-data-dir',
+			description: 'Specifies the directory that server data is kept in',
+			isRepeatable: true,
+			args: {
+				name: 'server_data_dir',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--extensions-dir',
+			description: 'Set the root path for extensions',
+			isRepeatable: true,
+			args: {
+				name: 'extensions_dir',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--tunnel-name',
+			description: 'Name you\'d like to assign preexisting tunnel to use to connect the tunnel Old option, new code should just use `--name`',
+			hidden: true,
+			isRepeatable: true,
+			args: {
+				name: 'tunnel_name',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--host-token',
+			description: 'Token to authenticate and use preexisting tunnel',
+			hidden: true,
+			isRepeatable: true,
+			args: {
+				name: 'host_token',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--tunnel-id',
+			description: 'ID of preexisting tunnel to use to connect the tunnel',
+			hidden: true,
+			isRepeatable: true,
+			args: {
+				name: 'tunnel_id',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--cluster',
+			description: 'Cluster of preexisting tunnel to use to connect the tunnel',
+			hidden: true,
+			isRepeatable: true,
+			args: {
+				name: 'cluster',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--name',
+			description: 'Sets the machine name for port forwarding service',
+			isRepeatable: true,
+			args: {
+				name: 'name',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--parent-process-id',
+			description: 'Optional parent process id. If provided, the server will be stopped when the process of the given pid no longer exists',
+			hidden: true,
+			isRepeatable: true,
+			args: {
+				name: 'parent_process_id',
+				isOptional: true,
+			},
+		},
+		{
+			name: '--random-name',
+			description: 'Randomly name machine for port forwarding service',
+		},
+		{
+			name: '--no-sleep',
+			description: 'Prevents the machine going to sleep while this command runs',
+		},
+		{
+			name: '--accept-server-license-terms',
+			description: 'If set, the user accepts the server license terms and the server will be started without a user prompt',
+		},
+		...commonCLIOptions,
+	],
+}];
 
 
 export default codeTunnelCompletionSpec;
